requirements.txt
streamlit
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-text-splitters==0.3.0
langchain-google-genai
faiss-cpu
pypdf
python-dotenv
notion-client
fpdf
tiktoken